remark: prim type to its wrap

+++ neobisss/src/Week4/src/main/java/org/example/onlineshop/service/ProductService.java
@@ -1,14 +1,14 @@
 package org.example.onlineshop.service;
 
+import org.example.onlineshop.entity.Product;
+import org.example.onlineshop.exception.EmptyFieldException;
 import lombok.AllArgsConstructor;
 import org.example.onlineshop.dto.ProductDTO;
-import org.example.onlineshop.entity.Product;
+import org.example.onlineshop.exception.ProductNotFoundException;
 import org.example.onlineshop.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,78 +20,61 @@ public class ProductService {
     public List<Product> getAllProducts(){
         return this.productRepository.findAll();
     }
-    public Optional<Product> findProductById(long id){
+    public Optional<Product> findProductById(Long id){
         return this.productRepository.findById(id);
     }
-    public  ResponseEntity<String> addProduct(String productName, double price){
+    public  Product addProduct(String productName, BigDecimal price){
         ProductDTO productDto = new ProductDTO();
-        Product product = new Product();
         productDto.setProductName(productName);
         productDto.setPrice(price);
-        if(productDto.getProductName().isEmpty() || productDto.getProductName().equals(" ")){
-            return ResponseEntity.badRequest().body("Product name is required!") ;
+        return addProductBody(productDto);
+    }
+    public Product addProductBody(ProductDTO productDTO){
+        if(productDTO.getProductName().isEmpty() || productDTO.getProductName().equals(" ")){
+            throw new EmptyFieldException("The field must not be empty!");
         }
-        else if(productDto.getPrice() == 0.0){
-            return ResponseEntity.badRequest().body("Product price is required!") ;
+        else if(productDTO.getPrice().equals(BigDecimal.valueOf(0))){
+            throw new EmptyFieldException("The field must not be empty!");
         }
         else{
-            product.setId(productDto.getId());
-            product.setProductName(productDto.getProductName());
-            product.setPrice(productDto.getPrice());
+            Product product = new Product();
+            product.setId(productDTO.getId());
+            product.setProductName(productDTO.getProductName());
+            product.setPrice(productDTO.getPrice());
             this.productRepository.save(product);
-            System.out.println("Added:  " + product.toString());
-            return ResponseEntity.ok("Added:  " + product.toString());
+            return product;
         }
     }
-    public ResponseEntity<String> addProductBody(ProductDTO productDto){
-        Product product = new Product();
-        if(productDto.getProductName().isEmpty() || productDto.getProductName().equals(" ")){
-            return ResponseEntity.badRequest().body("Product name is required!") ;
-        }
-        else if(productDto.getPrice() == 0.0){
-            return ResponseEntity.badRequest().body("Product price is required!") ;
-        }
-        else{
-            product.setId(productDto.getId());
-            product.setProductName(productDto.getProductName());
-            product.setPrice(productDto.getPrice());
-            this.productRepository.save(product);
-            System.out.println("Added:  " + product.toString());
-            return ResponseEntity.ok("Added:  " + product.toString());
-        }
-    }
-    public ResponseEntity<String> editProductById(long id, String productName, double price) {
-        Optional<Product> product = productRepository.findById(id);
+    public Product editProduct(ProductDTO productDTO) {
+        Optional<Product> product = productRepository.findById(productDTO.getId());
         if(product.isPresent()){
             Product updateProduct = product.get();
-            updateProduct.setProductName(productName);
-            updateProduct.setPrice(price);
+            updateProduct.setProductName(productDTO.getProductName());
+            updateProduct.setPrice(productDTO.getPrice());
             this.productRepository.save(updateProduct);
-            System.out.println("Updated! : "+updateProduct.toString());
-            return ResponseEntity.ok("Updated! : "+updateProduct.toString());
+            return updateProduct;
         }
         else{
-            System.out.println("Product not found for ID = " + id);
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body("Product not found for ID = " + id);
+            throw new ProductNotFoundException("Product not found!");
         }
     }
-    public ResponseEntity<String> deleteProductById(long id){
+    public String deleteProductById(Long id){
         Optional<Product> product = productRepository.findById(id);
         if(product.isPresent()){
             this.productRepository.deleteById(id);
-            System.out.println("Product is deleted with ID = "+id);
-            return ResponseEntity.ok("Product is deleted with ID = "+id);
+            return "Product deleted with id: " + id;
         }
         else{
-            System.out.println("Product not found for ID = " + id);
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body("Product not found for ID = " + id);
+            throw new ProductNotFoundException("Product not found!");
         }
     }
-    public String deleteAllProducts(){
-        this.productRepository.deleteAll();
-        System.out.println("All are deleted!");
-        return "All are deleted!";
+    public String deleteAllProducts()throws ProductNotFoundException{
+        if(this.productRepository.findAll().isEmpty()){
+            throw  new ProductNotFoundException("DB is already clear!");
+        }
+        else {
+            this.productRepository.deleteAll();
+            return "All are deleted!!!";
+        }
     }
 }
